requirements.txt
# ConstructAI Requirements

# Frontend
streamlit==1.35.0
Pillow==10.3.0

# Backend
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.1

# Vision AI
opencv-python-headless==4.9.0.80
ultralytics==8.2.0          # YOLOv8
numpy==1.26.4

# Database
supabase==2.4.6

# Utilities
python-dotenv==1.0.1
requests==2.32.2
httpx==0.27.0
